% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{has_char_length}
\alias{has_char_length}
\title{Check whether the variable has specified character length}
\usage{
has_char_length(x, .length, .min, .max)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a valid URL, FALSE otherwise
}
\description{
If a character vector is supplied each element is checked, returning a
logical vector of the same length.
}
\examples{
has_char_length("", 0)
has_char_length(c("some", "text"), 4)
has_char_length(c("some", "text"), .min = 1)
has_char_length(c("some", "text"), .max = 10)
has_char_length(c("some", "text"), .min = 1, .max = 10)

has_char_length("sometext", 4)
has_char_length(c("different", "text"), .min = 5)
has_char_length(c("different", "text"), .max = 5)
has_char_length(c("different", "text"), .min = 5, .max = 8)

}
