% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{has_length}
\alias{has_length}
\title{Check whether the variable has specified length}
\usage{
has_length(x, .length, .min, .max)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a valid URL, FALSE otherwise
}
\description{
Check whether the variable has specified length
}
\examples{
has_length(1, 1)
has_length(c("bob", "jane"), 2)
has_length(list(x = 1, y = 2, z = 3), 3)

has_length(1:3, 1)
has_length(c("bob", "jane"), 3)
has_length(list(x = 1, y = 2, z = 3), 5)

has_length(1:3, .min = 2)
has_length(1:3, .max = 10)

has_length("A", .min = 2)
has_length(LETTERS, .max = 10)

}
