% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_boolean}
\alias{is_boolean}
\title{Checks whether the variable is a logical vector of length 1}
\usage{
is_boolean(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a logical vector of length 1, FALSE otherwise
}
\description{
Checks whether the variable is a logical vector of length 1
}
