% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_character}
\alias{is_character}
\title{Checks whether the variable is a character vector}
\usage{
is_character(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a character vector, FALSE otherwise
}
\description{
Checks whether the variable is a character vector
}
