% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_data_frame}
\alias{is_data_frame}
\title{Checks whether the variable is a data frame}
\usage{
is_data_frame(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a data frame, FALSE otherwise
}
\description{
Checks whether the variable is a data frame
}
