% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_dir}
\alias{is_dir}
\title{Checks whether the variable is a path to an existing directory}
\usage{
is_dir(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a path to an existing directory, FALSE otherwise
}
\description{
Checks whether the variable is a path to an existing directory
}
\examples{
is_dir(tempdir())
is_dir("/does/not/exist")
is_dir(1)

}
