% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_file}
\alias{is_file}
\title{Checks whether the variable is a path to an existing file}
\usage{
is_file(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a path to an existing file, FALSE otherwise
}
\description{
Checks whether the variable is a path to an existing file
}
\examples{
tmpfile <- tempfile()
file.create(tmpfile)

is_file(tmpfile)

is_file("/does/not/exist.txt")
is_file(1)

}
