% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_function}
\alias{is_function}
\title{Checks whether the variable is a function}
\usage{
is_function(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a function, FALSE otherwise
}
\description{
Checks whether the variable is a function
}
