% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_integer}
\alias{is_integer}
\title{Checks whether the variable is a numeric vector of integers}
\usage{
is_integer(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a numeric vector of integers, FALSE otherwise
}
\description{
Checks whether the variable is a numeric vector of integers
}
