% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_integerish}
\alias{is_integerish}
\alias{is_scalar_integerish}
\alias{is_numeric}
\alias{is_scalar_numeric}
\alias{is_factor}
\alias{is_data_frame}
\title{Check the variable type}
\usage{
is_integerish(x)

is_scalar_integerish(x)

is_numeric(x)

is_scalar_numeric(x)

is_factor(x)

is_data_frame(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a valid URL, FALSE otherwise
}
\description{
Check the variable type
}
\examples{
is_integerish(1:3)
is_integerish(c(1.0, 2.0))
is_integerish(3.142)

is_scalar_integerish(1:3)
is_scalar_integerish(c(1.0, 2.0))
is_scalar_integerish(3.142)

is_numeric(1:3)
is_numeric(c(1.2, 2.4))
is_numeric("text")

is_scalar_numeric(1)
is_scalar_numeric(c(1.2, 2.4))
is_scalar_numeric("text")

is_factor(factor(c("a", "b", "a")))
is_factor(1:3)

is_data_frame(data.frame(a = 1:26, b = letters))
is_data_frame(list(a = 1:26, b = letters))

}
