% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_list}
\alias{is_list}
\title{Checks whether the variable is a list}
\usage{
is_list(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a list, FALSE otherwise
}
\description{
Checks whether the variable is a list
}
