% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_logical}
\alias{is_logical}
\title{Checks whether the variable is a logical vector}
\usage{
is_logical(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a logical vector, FALSE otherwise
}
\description{
Checks whether the variable is a logical vector
}
