% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_na}
\alias{is_na}
\title{Checks whether the variable is NA}
\usage{
is_na(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is NA, FALSE otherwise
}
\description{
Checks whether the variable is NA
}
\examples{
is_na(1)
is_na("foo")
is_na(NA)

is_na(c(1, NA))
is_na(c(NA, NA))

}
