% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_natural}
\alias{is_natural}
\title{Checks whether the variable is a natural number}
\usage{
is_natural(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a natural number, FALSE otherwise
}
\description{
Checks whether the variable is a natural number
}
