% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_null}
\alias{is_null}
\title{Checks whether the variable is NULL}
\usage{
is_null(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is NULL, FALSE otherwise
}
\description{
Checks whether the variable is NULL
}
