% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_number}
\alias{is_number}
\title{Checks whether the variable is a numeric vector of length 1}
\usage{
is_number(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a numeric vector of length 1, FALSE otherwise
}
\description{
Checks whether the variable is a numeric vector of length 1
}
