% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_numeric}
\alias{is_numeric}
\title{Checks whether the variable is a numeric vector}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a numeric vector, FALSE otherwise
}
\description{
Checks whether the variable is a numeric vector
}
