% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_readable}
\alias{is_readable}
\title{Checks whether the variable is a path to an existing, readable file or directory}
\usage{
is_readable(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a path to an existing, readable file or directory, FALSE otherwise
}
\description{
Checks whether the variable is a path to an existing, readable file or directory
}
\examples{
tmpfile <- tempfile()
file.create(tmpfile)

is_readable(tmpfile)

is_readable("/does/not/exist.txt")
is_readable(1)

}
