% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_scalar}
\alias{is_scalar}
\title{Checks whether the variable is an atomic vector of length 1}
\usage{
is_scalar(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is an atomic vector of length 1, FALSE otherwise
}
\description{
Checks whether the variable is an atomic vector of length 1
}
