% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_vector}
\alias{is_vector}
\title{Checks whether the variable is an atomic vector}
\usage{
is_vector(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is an atomic vector, FALSE otherwise
}
\description{
Checks whether the variable is an atomic vector
}
