% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_writeable}
\alias{is_writeable}
\title{Checks whether the variable is a path to an existing, writeable file or directory}
\usage{
is_writeable(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a path to an existing, writeable file or directory, FALSE otherwise
}
\description{
Checks whether the variable is a path to an existing, writeable file or directory
}
\examples{
tmpfile <- tempfile()
file.create(tmpfile)

is_writeable(tmpfile)

is_writeable("/does/not/exist.txt")
is_writeable(1)

}
