% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgr-package.R
\docType{package}
\name{msgr-package}
\alias{msgr}
\alias{msgr-package}
\title{msgr: Extends messages, warnings and errors by adding levels of messages 
    and the option to record them in a log file.}
\description{
Extends messages, warnings and errors by adding levels of
    messages and the option to record them in a log file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ChadGoymer/msgr}
  \item Report bugs at \url{https://github.com/ChadGoymer/msgr/issues}
}

}
\author{
\strong{Maintainer}: Chad Goymer \email{chad.goymer@gmail.com}

}
\keyword{internal}
