% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgr-package.R
\docType{package}
\name{msgr}
\alias{msgr}
\title{Extends messages, warnings and errors by adding levels and log files}
\description{
Provides new functions \code{\link[=info]{info()}}, \code{\link[=warn]{warn()}} and \code{\link[=error]{error()}}, similar to \code{\link[=message]{message()}},
\code{\link[=warning]{warning()}} and \code{\link[=stop]{stop()}} respectively. However, the new functions can have a \code{level}
associated with them, so that when executed the global level option determines whether
they are shown or not. This allows debug modes, outputting more information. The can also
output all messages to a log file.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/ChadGoymer/msgr}
\item Report bugs at \url{https://github.com/ChadGoymer/msgr/issues}
}
}
\author{
Chad Goymer \email{chad.goymer@gmail.com}
}
