% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try-catch.R
\name{with_msgr}
\alias{with_msgr}
\title{Convert messages, warnings and stops into msgr messages}
\usage{
with_msgr(
  expr,
  level = 1,
  warn_level = level,
  error_level = 1,
  msg_level = getOption("msgr.level"),
  msg_types = getOption("msgr.types"),
  log_path = getOption("msgr.log_path")
)
}
\arguments{
\item{expr}{(expression) The expression to execute.}

\item{level}{(integer, optional) The level of messages, from 1 to 10.
Default: 1.}

\item{warn_level}{(integer, optional) The level of warnings, from 1 to 10.
Default: 1.}

\item{error_level}{(integer, optional) The level of errors, from 1 to 10.
Default: 1.}

\item{msg_level}{(integer, optional) The maximum level of messages to output.
Default: set in the option \code{"msgr.level"}.}

\item{msg_types}{(character, optional) The type to write or display. Must
either NULL or one or more from "INFO", "WARNING" or "ERROR". Default: set
in the option \code{"msgr.types"}.}

\item{log_path}{(string, optional) The file path to the text log file. If set
to "", then no logs are written. Default: set in the option
\code{"msgr.log_path"}.}
}
\value{
Whatever the expression returns
}
\description{
This function allows you to execute an expression and all messages, warnings
and stops are converted to \code{\link[=info]{info()}}, \code{\link[=warn]{warn()}} and \code{\link[=error]{error()}}
respectively.
}
\details{
This function can also be used to change the \code{level} of message, so if a
function raises a message with level 1, it can be downgraded to level 2+.
Similarly it can also be used to modify the \code{msg_level}. For example, if the
current \code{msg_level} is set to 1, and a function raises a message with level
3, the message will not be displayed. If you want it displayed for this
function but not everywhere else, you can use \code{with_msgr()} and set the
\code{msg_level} to 3 for that call alone.
}
\examples{
\dontrun{

test_msgr <- function(x) {
  if (x == "msg") message("This is a message")
  if (x == "wrn") warning("This is a warning")
  if (x == "err") stop("This is an error")
}

with_msgr(test_msgr("msg"))
with_msgr(test_msgr("wrn"))
with_msgr(test_msgr("err"))

with_msgr(test_msgr("msg"), level = 2)
with_msgr(test_msgr("wrn"), level = 2)
with_msgr(test_msgr("err"), level = 2)

test_lvl <- function(x) {
  if (x == "msg") info("This is a message", level = 3)
  if (x == "wrn") warn("This is a warning", level = 3)
  if (x == "err") error("This is an error", level = 3)
}

with_msgr(test_lvl("msg"), msg_level = 3)
with_msgr(test_lvl("wrn"), msg_level = 3)
with_msgr(test_lvl("err"), msg_level = 3)

}

}
