% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_file}
\alias{is_file}
\alias{is_dir}
\alias{is_readable}
\alias{is_writeable}
\alias{is_url}
\title{Check file paths and directories}
\usage{
is_file(x)

is_dir(x)

is_readable(x)

is_writeable(x)

is_url(x)
}
\arguments{
\item{x}{(any) The object to test}
}
\value{
TRUE if x is a path to an existing directory, FALSE otherwise
}
\description{
These functions check whether the directories or files exist and whether the
current user has read or write permissions. If a character vector is supplied
each element is checked, returning a logical vector of the same length.
}
\examples{

is_dir(tempdir())
is_dir("/does/not/exist")

tmpfile <- tempfile()
file.create(tmpfile)

is_file(tmpfile)
is_file("/does/not/exist.txt")

is_readable(tmpfile)
is_readable("/does/not/exist.txt")

is_writeable(tmpfile)
is_writeable("/does/not/exist.txt")

is_url("http://something.com")
is_url("https://google.com")
is_url(c("http://something.com", "https://google.com"))

is_url(1)
is_url("foo")
is_url(NA)

}
