% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_in}
\alias{is_in}
\alias{is_in_range}
\title{Check whether elements of a variable are allowed (or not)}
\usage{
is_in(x, .values)

is_in_range(x, .min, .max)
}
\arguments{
\item{x}{(any) The object with elements to test}

\item{.values}{(any) The allowed (or not allowed) values.}

\item{.min}{(numeric, optional) The minimum allowed value for a range. If not
supplied, no minimum is set.}

\item{.max}{(numeric, optional) The maximum allowed value for a range. If not
supplied, no maximum is set.}
}
\value{
TRUE if elements in x are in y, FALSE otherwise
}
\description{
\code{is_in()} checks whether the elements of \code{x} are in a set of allowed values.
\code{is_in_range()} checks whether the elements of \code{x} are within a numeric
range.
If \code{x} is a character vector each element is checked, returning a logical
vector of the same length.
}
\examples{
is_in("a", letters)
is_in(c("a", "b", "c"), letters)

is_in(1, LETTERS)
is_in(1:2, LETTERS)

is_in_range(1, .min = 0)
is_in_range(c(1.1, 2.2, 3.3), .min = 0, .max = 10)

is_in_range(0, .min = 1)
is_in_range(c(1.1, 2.2, 3.3), .min = 2, .max = 3)

}
